refactor: Formatter의 extension을 Double의 extension으로 변경, formatToDecimal 메서드 구현
